isolating and specifying attack roll
